add beginnings of game code and new ASCIITexture code

+++ src/crovasshun/Updatable.java
@@ -2,5 +2,5 @@ package crovasshun;
 
 public interface Updatable {
 	
-	public void update(float deltaTime);
+	public void update(long deltaTime);
 }
